Major refactor for MQTT stream management
Before, there were two separate listeners for the primary MQTT stream (the one where PM, BC, etc. values are sent through). The main issue with this is that the two listeners handled things differently. While on the map, the charts on the dashboard would not recieve data and update, and while on the dashboard, the map would not be receiving the PM values.

By moving the primary MQTT stream into App.vue, we can have it running seamlessly on both pages and set it up to process data in a way that works both for the dashboard and the map. It also allows chart data to be updated without needing to mount or open any of the actual chart instances.

As of this commit, the live readout on the car icon is broken. This should be fixed soon.